+++ src/components/eventList/index.ts
@@ -1,0 +1,1 @@
+export { EventList } from './EventList';
